# rating_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/average-rating/{providerID}": {
            "get": {
                "description": "Get the average rating for a specific service provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ratings"
                ],
                "summary": "Get average rating",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider ID",
                        "name": "providerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AverageRatingResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorModel"
                        }
                    }
                }
            }
        },
        "/submit-rating": {
            "post": {
                "description": "Submit a new rating for a service provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ratings"
                ],
                "summary": "Submit a rating",
                "parameters": [
                    {
                        "description": "Rating",
                        "name": "rating",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RatingRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.RatingRequest": {
            "type": "object",
            "required": [
                "provider_id",
                "rating"
            ],
            "properties": {
                "provider_id": {
                    "description": "UUID and required validation",
                    "type": "string",
                    "example": "d1dcc1d6-f1ee-49e8-acf6-d7a5de7f4bea"
                },
                "rating": {
                    "description": "Rating should be between 1 and 5",
                    "type": "number",
                    "maximum": 5,
                    "minimum": 1
                }
            }
        },
        "dto.AverageRatingResponseDto": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                }
            }
        },
        "utils.ErrorModel": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {}
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}